requirements.txt
requests2>=2.16
websockets>=7.0
